+++ bin/skonfig
@@ -2,17 +2,15 @@
 
 if __name__ == "__main__":
     import os
-    import sys
-
     skonfig_bindir_pardir = os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
-
+    import sys
     if os.path.isfile(os.path.join(skonfig_bindir_pardir, "skonfig", "__init__.py")):
         sys.path.insert(0, skonfig_bindir_pardir)
-
     import skonfig
     skonfig.THIS_IS_SKONFIG = True
-
     try:
-        sys.exit(skonfig.run())
+        if skonfig.run():
+            sys.exit(0)
+        sys.exit(1)
     except KeyboardInterrupt:
         sys.exit(2)
